LayoutTests/imported/w3c/web-platform-tests/css/css-text-decor/animations/text-underline-offset-interpolation-expected.txt


PASS CSS Transitions: property <text-underline-offset> from [15px] to [0px] at (0) should be [15px]
PASS CSS Transitions: property <text-underline-offset> from [15px] to [0px] at (0.3) should be [10.5px]
PASS CSS Transitions: property <text-underline-offset> from [15px] to [0px] at (0.6) should be [6px]
PASS CSS Transitions: property <text-underline-offset> from [15px] to [0px] at (1) should be [0px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [15px] to [0px] at (0) should be [15px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [15px] to [0px] at (0.3) should be [10.5px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [15px] to [0px] at (0.6) should be [6px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [15px] to [0px] at (1) should be [0px]
PASS CSS Animations: property <text-underline-offset> from [15px] to [0px] at (0) should be [15px]
PASS CSS Animations: property <text-underline-offset> from [15px] to [0px] at (0.3) should be [10.5px]
PASS CSS Animations: property <text-underline-offset> from [15px] to [0px] at (0.6) should be [6px]
PASS CSS Animations: property <text-underline-offset> from [15px] to [0px] at (1) should be [0px]
PASS Web Animations: property <text-underline-offset> from [15px] to [0px] at (0) should be [15px]
PASS Web Animations: property <text-underline-offset> from [15px] to [0px] at (0.3) should be [10.5px]
PASS Web Animations: property <text-underline-offset> from [15px] to [0px] at (0.6) should be [6px]
PASS Web Animations: property <text-underline-offset> from [15px] to [0px] at (1) should be [0px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [0px] at (0) should be [16px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [0px] at (0.3) should be [11.2px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [0px] at (0.6) should be [6.4px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [0px] at (1) should be [0px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0px] at (0) should be [16px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0px] at (0.3) should be [11.2px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0px] at (0.6) should be [6.4px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0px] at (1) should be [0px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [0px] at (0) should be [16px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [0px] at (0.3) should be [11.2px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [0px] at (0.6) should be [6.4px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [0px] at (1) should be [0px]
PASS Web Animations: property <text-underline-offset> from [16px] to [0px] at (0) should be [16px]
PASS Web Animations: property <text-underline-offset> from [16px] to [0px] at (0.3) should be [11.2px]
PASS Web Animations: property <text-underline-offset> from [16px] to [0px] at (0.6) should be [6.4px]
PASS Web Animations: property <text-underline-offset> from [16px] to [0px] at (1) should be [0px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [32px] at (0) should be [16px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [32px] at (0.3) should be [20.8px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [32px] at (0.6) should be [25.6px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [32px] at (1) should be [32px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [32px] at (0) should be [16px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [32px] at (0.3) should be [20.8px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [32px] at (0.6) should be [25.6px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [32px] at (1) should be [32px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [32px] at (0) should be [16px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [32px] at (0.3) should be [20.8px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [32px] at (0.6) should be [25.6px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [32px] at (1) should be [32px]
PASS Web Animations: property <text-underline-offset> from [16px] to [32px] at (0) should be [16px]
PASS Web Animations: property <text-underline-offset> from [16px] to [32px] at (0.3) should be [20.8px]
PASS Web Animations: property <text-underline-offset> from [16px] to [32px] at (0.6) should be [25.6px]
PASS Web Animations: property <text-underline-offset> from [16px] to [32px] at (1) should be [32px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [0em] at (0) should be [16px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [0em] at (0.3) should be [11.2px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [0em] at (0.6) should be [6.4px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [0em] at (1) should be [0px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0em] at (0) should be [16px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0em] at (0.3) should be [11.2px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0em] at (0.6) should be [6.4px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0em] at (1) should be [0px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [0em] at (0) should be [16px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [0em] at (0.3) should be [11.2px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [0em] at (0.6) should be [6.4px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [0em] at (1) should be [0px]
PASS Web Animations: property <text-underline-offset> from [1em] to [0em] at (0) should be [16px]
PASS Web Animations: property <text-underline-offset> from [1em] to [0em] at (0.3) should be [11.2px]
PASS Web Animations: property <text-underline-offset> from [1em] to [0em] at (0.6) should be [6.4px]
PASS Web Animations: property <text-underline-offset> from [1em] to [0em] at (1) should be [0px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [2em] at (0) should be [16px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [2em] at (0.3) should be [20.8px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [2em] at (0.6) should be [25.6px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [2em] at (1) should be [32px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [2em] at (0) should be [16px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [2em] at (0.3) should be [20.8px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [2em] at (0.6) should be [25.6px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [2em] at (1) should be [32px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [2em] at (0) should be [16px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [2em] at (0.3) should be [20.8px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [2em] at (0.6) should be [25.6px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [2em] at (1) should be [32px]
PASS Web Animations: property <text-underline-offset> from [1em] to [2em] at (0) should be [16px]
PASS Web Animations: property <text-underline-offset> from [1em] to [2em] at (0.3) should be [20.8px]
PASS Web Animations: property <text-underline-offset> from [1em] to [2em] at (0.6) should be [25.6px]
PASS Web Animations: property <text-underline-offset> from [1em] to [2em] at (1) should be [32px]
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0%] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0%] at (0.3) should be [70%] assert_equals: expected "70 % " but got "11.2px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0%] at (0.6) should be [40%] assert_equals: expected "40 % " but got "6.4px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0%] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0%] at (0.3) should be [70%] assert_equals: expected "70 % " but got "11.2px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0%] at (0.6) should be [40%] assert_equals: expected "40 % " but got "6.4px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0%] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0%] at (0.3) should be [70%] assert_equals: expected "70 % " but got "11.2px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0%] at (0.6) should be [40%] assert_equals: expected "40 % " but got "6.4px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0%] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0%] at (0.3) should be [70%] assert_equals: expected "70 % " but got "11.2px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0%] at (0.6) should be [40%] assert_equals: expected "40 % " but got "6.4px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [200%] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [200%] at (0.3) should be [130%] assert_equals: expected "130 % " but got "20.8px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [200%] at (0.6) should be [160%] assert_equals: expected "160 % " but got "25.6px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [200%] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [200%] at (0.3) should be [130%] assert_equals: expected "130 % " but got "20.8px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [200%] at (0.6) should be [160%] assert_equals: expected "160 % " but got "25.6px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [200%] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [200%] at (0.3) should be [130%] assert_equals: expected "130 % " but got "20.8px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [200%] at (0.6) should be [160%] assert_equals: expected "160 % " but got "25.6px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [200%] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [200%] at (0.3) should be [130%] assert_equals: expected "130 % " but got "20.8px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [200%] at (0.6) should be [160%] assert_equals: expected "160 % " but got "25.6px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
PASS CSS Transitions: property <text-underline-offset> from [16px] to [0em] at (0) should be [16px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [0em] at (0.3) should be [11.2px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [0em] at (0.6) should be [6.4px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [0em] at (1) should be [0px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0em] at (0) should be [16px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0em] at (0.3) should be [11.2px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0em] at (0.6) should be [6.4px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0em] at (1) should be [0px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [0em] at (0) should be [16px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [0em] at (0.3) should be [11.2px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [0em] at (0.6) should be [6.4px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [0em] at (1) should be [0px]
PASS Web Animations: property <text-underline-offset> from [16px] to [0em] at (0) should be [16px]
PASS Web Animations: property <text-underline-offset> from [16px] to [0em] at (0.3) should be [11.2px]
PASS Web Animations: property <text-underline-offset> from [16px] to [0em] at (0.6) should be [6.4px]
PASS Web Animations: property <text-underline-offset> from [16px] to [0em] at (1) should be [0px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [2em] at (0) should be [16px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [2em] at (0.3) should be [20.8px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [2em] at (0.6) should be [25.6px]
PASS CSS Transitions: property <text-underline-offset> from [16px] to [2em] at (1) should be [32px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [2em] at (0) should be [16px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [2em] at (0.3) should be [20.8px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [2em] at (0.6) should be [25.6px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [2em] at (1) should be [32px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [2em] at (0) should be [16px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [2em] at (0.3) should be [20.8px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [2em] at (0.6) should be [25.6px]
PASS CSS Animations: property <text-underline-offset> from [16px] to [2em] at (1) should be [32px]
PASS Web Animations: property <text-underline-offset> from [16px] to [2em] at (0) should be [16px]
PASS Web Animations: property <text-underline-offset> from [16px] to [2em] at (0.3) should be [20.8px]
PASS Web Animations: property <text-underline-offset> from [16px] to [2em] at (0.6) should be [25.6px]
PASS Web Animations: property <text-underline-offset> from [16px] to [2em] at (1) should be [32px]
FAIL CSS Transitions: property <text-underline-offset> from [16px] to [0%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Transitions: property <text-underline-offset> from [16px] to [0%] at (0.3) should be [calc(0% + 11.2px)] assert_equals: expected "calc ( 0 % + 11.2px ) " but got "11.2px "
FAIL CSS Transitions: property <text-underline-offset> from [16px] to [0%] at (0.6) should be [calc(0% + 6.4px)] assert_equals: expected "calc ( 0 % + 6.4px ) " but got "6.4px "
FAIL CSS Transitions: property <text-underline-offset> from [16px] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0%] at (0.3) should be [calc(0% + 11.2px)] assert_equals: expected "calc ( 0 % + 11.2px ) " but got "11.2px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0%] at (0.6) should be [calc(0% + 6.4px)] assert_equals: expected "calc ( 0 % + 6.4px ) " but got "6.4px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Animations: property <text-underline-offset> from [16px] to [0%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [16px] to [0%] at (0.3) should be [calc(0% + 11.2px)] assert_equals: expected "calc ( 0 % + 11.2px ) " but got "11.2px "
FAIL CSS Animations: property <text-underline-offset> from [16px] to [0%] at (0.6) should be [calc(0% + 6.4px)] assert_equals: expected "calc ( 0 % + 6.4px ) " but got "6.4px "
FAIL CSS Animations: property <text-underline-offset> from [16px] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL Web Animations: property <text-underline-offset> from [16px] to [0%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [16px] to [0%] at (0.3) should be [calc(0% + 11.2px)] assert_equals: expected "calc ( 0 % + 11.2px ) " but got "11.2px "
FAIL Web Animations: property <text-underline-offset> from [16px] to [0%] at (0.6) should be [calc(0% + 6.4px)] assert_equals: expected "calc ( 0 % + 6.4px ) " but got "6.4px "
FAIL Web Animations: property <text-underline-offset> from [16px] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions: property <text-underline-offset> from [16px] to [200%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Transitions: property <text-underline-offset> from [16px] to [200%] at (0.3) should be [calc(120% + 11.2px)] assert_equals: expected "calc ( 120 % + 11.2px ) " but got "20.8px "
FAIL CSS Transitions: property <text-underline-offset> from [16px] to [200%] at (0.6) should be [calc(120% + 6.4px)] assert_equals: expected "calc ( 120 % + 6.4px ) " but got "25.6px "
FAIL CSS Transitions: property <text-underline-offset> from [16px] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [200%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [200%] at (0.3) should be [calc(120% + 11.2px)] assert_equals: expected "calc ( 120 % + 11.2px ) " but got "20.8px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [200%] at (0.6) should be [calc(120% + 6.4px)] assert_equals: expected "calc ( 120 % + 6.4px ) " but got "25.6px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [16px] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
FAIL CSS Animations: property <text-underline-offset> from [16px] to [200%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [16px] to [200%] at (0.3) should be [calc(120% + 11.2px)] assert_equals: expected "calc ( 120 % + 11.2px ) " but got "20.8px "
FAIL CSS Animations: property <text-underline-offset> from [16px] to [200%] at (0.6) should be [calc(120% + 6.4px)] assert_equals: expected "calc ( 120 % + 6.4px ) " but got "25.6px "
FAIL CSS Animations: property <text-underline-offset> from [16px] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
FAIL Web Animations: property <text-underline-offset> from [16px] to [200%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [16px] to [200%] at (0.3) should be [calc(120% + 11.2px)] assert_equals: expected "calc ( 120 % + 11.2px ) " but got "20.8px "
FAIL Web Animations: property <text-underline-offset> from [16px] to [200%] at (0.6) should be [calc(120% + 6.4px)] assert_equals: expected "calc ( 120 % + 6.4px ) " but got "25.6px "
FAIL Web Animations: property <text-underline-offset> from [16px] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
PASS CSS Transitions: property <text-underline-offset> from [1em] to [0px] at (0) should be [16px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [0px] at (0.3) should be [11.2px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [0px] at (0.6) should be [6.4px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [0px] at (1) should be [0px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0px] at (0) should be [16px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0px] at (0.3) should be [11.2px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0px] at (0.6) should be [6.4px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0px] at (1) should be [0px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [0px] at (0) should be [16px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [0px] at (0.3) should be [11.2px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [0px] at (0.6) should be [6.4px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [0px] at (1) should be [0px]
PASS Web Animations: property <text-underline-offset> from [1em] to [0px] at (0) should be [16px]
PASS Web Animations: property <text-underline-offset> from [1em] to [0px] at (0.3) should be [11.2px]
PASS Web Animations: property <text-underline-offset> from [1em] to [0px] at (0.6) should be [6.4px]
PASS Web Animations: property <text-underline-offset> from [1em] to [0px] at (1) should be [0px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [32px] at (0) should be [16px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [32px] at (0.3) should be [20.8px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [32px] at (0.6) should be [25.6px]
PASS CSS Transitions: property <text-underline-offset> from [1em] to [32px] at (1) should be [32px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [32px] at (0) should be [16px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [32px] at (0.3) should be [20.8px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [32px] at (0.6) should be [25.6px]
PASS CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [32px] at (1) should be [32px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [32px] at (0) should be [16px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [32px] at (0.3) should be [20.8px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [32px] at (0.6) should be [25.6px]
PASS CSS Animations: property <text-underline-offset> from [1em] to [32px] at (1) should be [32px]
PASS Web Animations: property <text-underline-offset> from [1em] to [32px] at (0) should be [16px]
PASS Web Animations: property <text-underline-offset> from [1em] to [32px] at (0.3) should be [20.8px]
PASS Web Animations: property <text-underline-offset> from [1em] to [32px] at (0.6) should be [25.6px]
PASS Web Animations: property <text-underline-offset> from [1em] to [32px] at (1) should be [32px]
FAIL CSS Transitions: property <text-underline-offset> from [1em] to [0%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Transitions: property <text-underline-offset> from [1em] to [0%] at (0.3) should be [calc(0% + 11.2px)] assert_equals: expected "calc ( 0 % + 11.2px ) " but got "11.2px "
FAIL CSS Transitions: property <text-underline-offset> from [1em] to [0%] at (0.6) should be [calc(0% + 6.4px)] assert_equals: expected "calc ( 0 % + 6.4px ) " but got "6.4px "
FAIL CSS Transitions: property <text-underline-offset> from [1em] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0%] at (0.3) should be [calc(0% + 11.2px)] assert_equals: expected "calc ( 0 % + 11.2px ) " but got "11.2px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0%] at (0.6) should be [calc(0% + 6.4px)] assert_equals: expected "calc ( 0 % + 6.4px ) " but got "6.4px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Animations: property <text-underline-offset> from [1em] to [0%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [1em] to [0%] at (0.3) should be [calc(0% + 11.2px)] assert_equals: expected "calc ( 0 % + 11.2px ) " but got "11.2px "
FAIL CSS Animations: property <text-underline-offset> from [1em] to [0%] at (0.6) should be [calc(0% + 6.4px)] assert_equals: expected "calc ( 0 % + 6.4px ) " but got "6.4px "
FAIL CSS Animations: property <text-underline-offset> from [1em] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL Web Animations: property <text-underline-offset> from [1em] to [0%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [1em] to [0%] at (0.3) should be [calc(0% + 11.2px)] assert_equals: expected "calc ( 0 % + 11.2px ) " but got "11.2px "
FAIL Web Animations: property <text-underline-offset> from [1em] to [0%] at (0.6) should be [calc(0% + 6.4px)] assert_equals: expected "calc ( 0 % + 6.4px ) " but got "6.4px "
FAIL Web Animations: property <text-underline-offset> from [1em] to [0%] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions: property <text-underline-offset> from [1em] to [200%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Transitions: property <text-underline-offset> from [1em] to [200%] at (0.3) should be [calc(60% + 11.2px)] assert_equals: expected "calc ( 60 % + 11.2px ) " but got "20.8px "
FAIL CSS Transitions: property <text-underline-offset> from [1em] to [200%] at (0.6) should be [calc(120% + 6.4px)] assert_equals: expected "calc ( 120 % + 6.4px ) " but got "25.6px "
FAIL CSS Transitions: property <text-underline-offset> from [1em] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [200%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [200%] at (0.3) should be [calc(60% + 11.2px)] assert_equals: expected "calc ( 60 % + 11.2px ) " but got "20.8px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [200%] at (0.6) should be [calc(120% + 6.4px)] assert_equals: expected "calc ( 120 % + 6.4px ) " but got "25.6px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [1em] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
FAIL CSS Animations: property <text-underline-offset> from [1em] to [200%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [1em] to [200%] at (0.3) should be [calc(60% + 11.2px)] assert_equals: expected "calc ( 60 % + 11.2px ) " but got "20.8px "
FAIL CSS Animations: property <text-underline-offset> from [1em] to [200%] at (0.6) should be [calc(120% + 6.4px)] assert_equals: expected "calc ( 120 % + 6.4px ) " but got "25.6px "
FAIL CSS Animations: property <text-underline-offset> from [1em] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
FAIL Web Animations: property <text-underline-offset> from [1em] to [200%] at (0) should be [calc(0% + 16px)] assert_equals: expected "calc ( 0 % + 16px ) " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [1em] to [200%] at (0.3) should be [calc(60% + 11.2px)] assert_equals: expected "calc ( 60 % + 11.2px ) " but got "20.8px "
FAIL Web Animations: property <text-underline-offset> from [1em] to [200%] at (0.6) should be [calc(120% + 6.4px)] assert_equals: expected "calc ( 120 % + 6.4px ) " but got "25.6px "
FAIL Web Animations: property <text-underline-offset> from [1em] to [200%] at (1) should be [200%] assert_equals: expected "200 % " but got "32px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0px] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0px] at (0.3) should be [70%] assert_equals: expected "70 % " but got "11.2px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0px] at (0.6) should be [40%] assert_equals: expected "40 % " but got "6.4px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0px] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0px] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0px] at (0.3) should be [70%] assert_equals: expected "70 % " but got "11.2px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0px] at (0.6) should be [40%] assert_equals: expected "40 % " but got "6.4px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0px] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0px] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0px] at (0.3) should be [70%] assert_equals: expected "70 % " but got "11.2px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0px] at (0.6) should be [40%] assert_equals: expected "40 % " but got "6.4px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0px] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0px] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0px] at (0.3) should be [70%] assert_equals: expected "70 % " but got "11.2px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0px] at (0.6) should be [40%] assert_equals: expected "40 % " but got "6.4px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0px] at (1) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [32px] at (0) should be [calc(100% + 0px)] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [32px] at (0.3) should be [calc(100% + 9.6px)] assert_equals: expected "calc ( 100 % + 9.6px ) " but got "20.8px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [32px] at (0.6) should be [calc(100% + 19.2px)] assert_equals: expected "calc ( 100 % + 19.2px ) " but got "25.6px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [32px] at (1) should be [calc(100% + 32px)] assert_equals: expected "calc ( 100 % + 32px ) " but got "32px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [32px] at (0) should be [calc(100% + 0px)] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [32px] at (0.3) should be [calc(100% + 9.6px)] assert_equals: expected "calc ( 100 % + 9.6px ) " but got "20.8px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [32px] at (0.6) should be [calc(100% + 19.2px)] assert_equals: expected "calc ( 100 % + 19.2px ) " but got "25.6px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [32px] at (1) should be [calc(100% + 32px)] assert_equals: expected "calc ( 100 % + 32px ) " but got "32px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [32px] at (0) should be [calc(100% + 0px)] assert_equals: expected "100 % " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [32px] at (0.3) should be [calc(100% + 9.6px)] assert_equals: expected "calc ( 100 % + 9.6px ) " but got "20.8px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [32px] at (0.6) should be [calc(100% + 19.2px)] assert_equals: expected "calc ( 100 % + 19.2px ) " but got "25.6px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [32px] at (1) should be [calc(100% + 32px)] assert_equals: expected "calc ( 100 % + 32px ) " but got "32px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [32px] at (0) should be [calc(100% + 0px)] assert_equals: expected "100 % " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [32px] at (0.3) should be [calc(100% + 9.6px)] assert_equals: expected "calc ( 100 % + 9.6px ) " but got "20.8px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [32px] at (0.6) should be [calc(100% + 19.2px)] assert_equals: expected "calc ( 100 % + 19.2px ) " but got "25.6px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [32px] at (1) should be [calc(100% + 32px)] assert_equals: expected "calc ( 100 % + 32px ) " but got "32px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0em] at (0) should be [calc(100% + 0em)] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0em] at (0.3) should be [calc(70% + 0em)] assert_equals: expected "70 % " but got "11.2px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0em] at (0.6) should be [calc(40% + 0em)] assert_equals: expected "40 % " but got "6.4px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [0em] at (1) should be [calc(0% + 0em)] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0em] at (0) should be [calc(100% + 0em)] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0em] at (0.3) should be [calc(70% + 0em)] assert_equals: expected "70 % " but got "11.2px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0em] at (0.6) should be [calc(40% + 0em)] assert_equals: expected "40 % " but got "6.4px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [0em] at (1) should be [calc(0% + 0em)] assert_equals: expected "0 % " but got "0px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0em] at (0) should be [calc(100% + 0em)] assert_equals: expected "100 % " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0em] at (0.3) should be [calc(70% + 0em)] assert_equals: expected "70 % " but got "11.2px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0em] at (0.6) should be [calc(40% + 0em)] assert_equals: expected "40 % " but got "6.4px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [0em] at (1) should be [calc(0% + 0em)] assert_equals: expected "0 % " but got "0px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0em] at (0) should be [calc(100% + 0em)] assert_equals: expected "100 % " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0em] at (0.3) should be [calc(70% + 0em)] assert_equals: expected "70 % " but got "11.2px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0em] at (0.6) should be [calc(40% + 0em)] assert_equals: expected "40 % " but got "6.4px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [0em] at (1) should be [calc(0% + 0em)] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [2em] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [2em] at (0.3) should be [calc(70% + 9.6px)] assert_equals: expected "calc ( 70 % + 9.6px ) " but got "20.8px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [2em] at (0.6) should be [calc(40% + 19.2px)] assert_equals: expected "calc ( 40 % + 19.2px ) " but got "25.6px "
FAIL CSS Transitions: property <text-underline-offset> from [100%] to [2em] at (1) should be [calc(0% + 32px)] assert_equals: expected "calc ( 0 % + 32px ) " but got "32px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [2em] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [2em] at (0.3) should be [calc(70% + 9.6px)] assert_equals: expected "calc ( 70 % + 9.6px ) " but got "20.8px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [2em] at (0.6) should be [calc(40% + 19.2px)] assert_equals: expected "calc ( 40 % + 19.2px ) " but got "25.6px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [100%] to [2em] at (1) should be [calc(0% + 32px)] assert_equals: expected "calc ( 0 % + 32px ) " but got "32px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [2em] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [2em] at (0.3) should be [calc(70% + 9.6px)] assert_equals: expected "calc ( 70 % + 9.6px ) " but got "20.8px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [2em] at (0.6) should be [calc(40% + 19.2px)] assert_equals: expected "calc ( 40 % + 19.2px ) " but got "25.6px "
FAIL CSS Animations: property <text-underline-offset> from [100%] to [2em] at (1) should be [calc(0% + 32px)] assert_equals: expected "calc ( 0 % + 32px ) " but got "32px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [2em] at (0) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [2em] at (0.3) should be [calc(70% + 9.6px)] assert_equals: expected "calc ( 70 % + 9.6px ) " but got "20.8px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [2em] at (0.6) should be [calc(40% + 19.2px)] assert_equals: expected "calc ( 40 % + 19.2px ) " but got "25.6px "
FAIL Web Animations: property <text-underline-offset> from [100%] to [2em] at (1) should be [calc(0% + 32px)] assert_equals: expected "calc ( 0 % + 32px ) " but got "32px "
FAIL CSS Transitions: property <text-underline-offset> from [0%] to [100%] at (0) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions: property <text-underline-offset> from [0%] to [100%] at (0.3) should be [30%] assert_equals: expected "30 % " but got "4.8px "
FAIL CSS Transitions: property <text-underline-offset> from [0%] to [100%] at (0.6) should be [60%] assert_equals: expected "60 % " but got "9.6px "
FAIL CSS Transitions: property <text-underline-offset> from [0%] to [100%] at (1) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [0%] to [100%] at (0) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [0%] to [100%] at (0.3) should be [30%] assert_equals: expected "30 % " but got "4.8px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [0%] to [100%] at (0.6) should be [60%] assert_equals: expected "60 % " but got "9.6px "
FAIL CSS Transitions with transition: all: property <text-underline-offset> from [0%] to [100%] at (1) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL CSS Animations: property <text-underline-offset> from [0%] to [100%] at (0) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Animations: property <text-underline-offset> from [0%] to [100%] at (0.3) should be [30%] assert_equals: expected "30 % " but got "4.8px "
FAIL CSS Animations: property <text-underline-offset> from [0%] to [100%] at (0.6) should be [60%] assert_equals: expected "60 % " but got "9.6px "
FAIL CSS Animations: property <text-underline-offset> from [0%] to [100%] at (1) should be [100%] assert_equals: expected "100 % " but got "16px "
FAIL Web Animations: property <text-underline-offset> from [0%] to [100%] at (0) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL Web Animations: property <text-underline-offset> from [0%] to [100%] at (0.3) should be [30%] assert_equals: expected "30 % " but got "4.8px "
FAIL Web Animations: property <text-underline-offset> from [0%] to [100%] at (0.6) should be [60%] assert_equals: expected "60 % " but got "9.6px "
FAIL Web Animations: property <text-underline-offset> from [0%] to [100%] at (1) should be [100%] assert_equals: expected "100 % " but got "16px "
